Increase default read timeout to 2s
This now matches HID_TIMEOUT from libu2f-host.

Empirically the Plug-Up U2F device takes around 1.05s
to respond to a enrollment, so fell foul of this timeout.  In
contrast the Yubico takes 0.14s so is comfortably within this
window.

At the same time, remove the unused TIMEOUT variable here.